requirements.txt
Flask
sqlalchemy
schedule==1.2.0
python-dotenv==1.0.0
google-generativeai
openai
fastapi
uvicorn
